foratting and todo fixes

+++ mp_mazes/dsets.cpp
@@ -9,11 +9,11 @@ void DisjointSets::addelements(uint32_t num) {
   _data.insert(_data.end(), num, -1);
 }
 
-int DisjointSets::find(uint32_t elem) {
+uint32_t DisjointSets::find(uint32_t elem) {
   if (_data[elem] < 0) {
     return elem;
   }
-  _data[elem] = find(_data[elem]);
+  _data[elem] = static_cast<int32_t>(find(_data[elem]));
   return _data[elem];
 }
 
@@ -23,11 +23,11 @@ void DisjointSets::setunion(uint32_t a, uint32_t b) {
 
   int new_size = _data[a] + _data[b];
   if (_data[a] < _data[b]) {
-    _data[b] = a;
+    _data[b] = static_cast<int32_t>(a);
     _data[a] = new_size;
     return;
   }
-  _data[a] = b;
+  _data[a] = static_cast<int32_t>(b);
   _data[b] = new_size;
 }
